fix(nord-cli): bound device controls --len to a transfer's wLength
`--len` was a usize that reached `vec![0u8; len]` before the request was
sent, so `--len 1e12` allocated rather than failing. A control transfer
states its length in 16 bits, so the flag now does too and clap refuses
anything wider.

+++ crates/nord-cli/src/main.rs
@@ -168,9 +168,9 @@ enum DeviceAction {
         #[arg(long, default_value_t = 15)]
         to: u8,
 
-        /// Bytes to ask each request for.
+        /// Bytes to ask each request for. A control transfer's wLength is 16 bits.
         #[arg(long, default_value_t = 64)]
-        len: usize,
+        len: u16,
 
         /// Address the interface rather than the device.
         #[arg(long)]
